+++ CHANGELOG.md
@@ -5,10 +5,12 @@ v6.4-SNAPSHOT
 ===
 - Analyzer
   - Added WebviewAppNameVersion.
+  - Added a 'not contains' operator '!~'
 - New/improved detections
   - More Apple devices
   - VivoBrowser often implies Vivo DeviceBrand
   - Handle the SalesforceMobileSDK apps better.
+  - More Amazon devices
 - Updated UDF dependencies
   - Apache Beam             2.34.0
   - Elastic Search/Logstash 7.15.0